+++ gplugins/sax/models.py
@@ -367,7 +367,7 @@ def _mmi_nxn(
                 splitting_matrix[i][j]
             )  # Convert power ratio to amplitude
             loss_factor = 10 ** (-loss_dB[j] / 20)
-            S[(f"o{i+1}", f"o{j+1}")] = amplitude * loss_factor
+            S[(f"o{i + 1}", f"o{j + 1}")] = amplitude * loss_factor
 
     return sax.reciprocal(S)
 
@@ -413,10 +413,9 @@ def mmi2x2(
         loss_dB: loss in dB.
         shift: shift in wavelength for both cross and thru ports.
         loss_dB_cross: loss in dB for the cross port.
-        loss_dB_bar: loss in dB for the bar port.
+        loss_dB_thru: loss in dB for the bar port.
         splitting_ratio_cross: splitting ratio for the cross port.
         splitting_ratio_thru: splitting ratio for the bar port.
-
     """
     loss_dB_cross = loss_dB_cross or loss_dB
     loss_dB_thru = loss_dB_thru or loss_dB