+++ gplugins/typings.py
@@ -1,0 +1,6 @@
+from typing import Any
+
+import numpy as np
+import numpy.typing as npt
+
+NDArrayF = npt.NDArray[np.floating[Any]]